+++ src/utilities/sessionUtils.js
@@ -1,9 +1,9 @@
-import {formatDateTime} from './utils';
+import {formatDateTimeISO} from './utils';
 
 export function getImageRecords(target) {
   return target.imageRecords.map((row) => ({
     ...row,
-    started: formatDateTime(row.started),
+    started: formatDateTimeISO(row.started),
     hfrText: Number.parseFloat(row.HFR).toPrecision(4),
   }));
 }
@@ -16,4 +16,3 @@ export function getThumbnailSize(sessionPath, item, notify) {
 
   img.src = sessionPath + '/thumbnails/' + item.id + '.jpg';
 }
-