+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # NINA Web Client
 
-## 1.0.0.0 - 2022-01-dd
+## 1.1.0.0 - 2022-04-dd
+* Added timeline plot of NINA events including all images saved, AF, MF, and more
+* Hover over events to see details (image thumbnails, AF curve plots)
+* Added plot of image quality metrics: stars, HFR, ADU stats, guiding stats, temperature
+
+## 1.0.0.0 - 2022-01-21
 * Initial release